Add parent class MrozaViewPage with similar for every page methods
