+++ MROZA-Show/MROZA-Show-web/src/test/java/com/mroza/seleniumTests/EditProgramViewTests/EditProgramViewPage.java
@@ -18,6 +18,7 @@
 package com.mroza.seleniumTests.EditProgramViewTests;
 
 import com.mroza.models.Program;
+import com.mroza.seleniumTests.MrozaViewPage;
 import com.mroza.utils.SeleniumWaiter;
 import com.mroza.utils.Utils;
 import org.openqa.selenium.By;
@@ -27,23 +28,16 @@ import org.openqa.selenium.WebElement;
 import java.util.ArrayList;
 import java.util.List;
 
-public class EditProgramViewPage {
-
-    private final WebDriver driver;
+public class EditProgramViewPage extends MrozaViewPage{
 
     public EditProgramViewPage(WebDriver driver) {
-        this.driver = driver;
+        super(driver);
     }
     public void open(String url, Program program) {
         driver.get(url + program.getId());
     }
     public void close() {
         driver.quit();
-    }
-
-    public String getHeader() {
-        WebElement header = driver.findElement(By.className("b-page-header"));
-        return header.getText();
     }
 
     public List<String> getTablesNamesListContent() {
@@ -67,24 +61,9 @@ public class EditProgramViewPage {
     }
 
     public void clickAddNewTableButton() {
-        clickButtonNamed(Utils.getMsgFromResources("editProgramView.addTable"));
+        clickButtonInButtonContainerNamed(Utils.getMsgFromResources("editProgramView.addTable"));
     }
 
-    private void clickButtonNamed(String buttonName) {
-        WebElement buttonsArea = driver.findElement(By.className("action-buttons-container"));
-        List<WebElement> buttonsList = buttonsArea.findElements(By.tagName("button"));
-
-        for(WebElement button : buttonsList)
-        {
-            if(button.findElement(By.tagName("span")).getText().equals(buttonName))
-            {
-                button.click();
-                SeleniumWaiter.waitForJQueryAndPrimeFaces(driver);
-                break;
-            }
-
-        }
-    }
 
     public void inputTableName(String tableName) {
         WebElement table = getTableWithHeader("", true);
@@ -94,8 +73,8 @@ public class EditProgramViewPage {
     }
 
 
-    public void changeTableName(String tableName, String oldTbleName) {
-        WebElement table = getTableWithHeader(oldTbleName, true);
+    public void changeTableName(String tableName, String oldTableName) {
+        WebElement table = getTableWithHeader(oldTableName, true);
         WebElement tableHeader = table.findElement(By.className("ui-datatable-header"));
         tableHeader.findElement(By.tagName("input")).clear();
         tableHeader.findElement(By.tagName("input")).sendKeys(tableName);
@@ -116,13 +95,7 @@ public class EditProgramViewPage {
 
     private void clickButtonForTable(String tableName, String buttonName, Boolean isEditing) {
         WebElement table = getTableWithHeader(tableName, isEditing);
-        List<WebElement> tableButtons = table.findElements(By.tagName("button"));
-        for(WebElement button : tableButtons)
-            if(button.findElement(By.tagName("span")).getText().equals(buttonName)) {
-                button.click();
-                SeleniumWaiter.waitForJQueryAndPrimeFaces(driver);
-                break;
-            }
+        clickButtonIn(buttonName,table);
     }
 
     public void changeTableRow(String tableName, String rowName, String oldRowName) {
@@ -180,18 +153,6 @@ public class EditProgramViewPage {
         return tableRowNames;
     }
 
-    public String getErrorMessage()
-    {
-        try{
-            WebElement messageArea = driver.findElement(By.className("ui-messages-error-summary"));
-            return messageArea.getText();
-        }
-        catch (Exception e)
-        {
-            return "MESSAGE NOT FOUND";
-        }
-
-    }
 
 
     private List<WebElement> getTablesList() {
