+++ MROZA-Show/MROZA-Show-web/src/test/java/com/mroza/seleniumTests/KidProgramsViewTests/KidProgramsViewPage.java
@@ -18,6 +18,7 @@
 package com.mroza.seleniumTests.KidProgramsViewTests;
 
 import com.mroza.models.Kid;
+import com.mroza.seleniumTests.MrozaViewPage;
 import com.mroza.utils.SeleniumWaiter;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
@@ -26,12 +27,11 @@ import org.openqa.selenium.WebElement;
 import java.util.ArrayList;
 import java.util.List;
 
-public class KidProgramsViewPage {
+public class KidProgramsViewPage extends MrozaViewPage{
 
-    protected WebDriver driver;
 
     public KidProgramsViewPage(WebDriver driver) {
-        this.driver = driver;
+        super(driver);
     }
     public void open(String url, Kid kid) {
         driver.get(url + kid.getId() + "&reloadData=true");
@@ -39,10 +39,7 @@ public class KidProgramsViewPage {
     public void close() {
         driver.quit();
     }
-    public String getHeader() {
-        WebElement header = driver.findElement(By.className("b-page-header"));
-        return header.getText();
-    }
+
 
     public void setSearchValue(String symbol){
         WebElement searchBoxInput = driver.findElement(By.className("ui-inputtext"));
